Object Avoiding Car
This code needs a few libraries which can be downloaded from the given links in the code. This is a 4 wheel drive car with an Arduino, Arduino motor driver shield, with a servo motor and an ultrasonic sensor. When uploaded this code lets the car run autonomously without any collison.