Voice controlled car with obstacle avoidance
This code needs a few libraries which can be downloaded from the given links in the code. This is a 4 wheel drive car with an Arduino, Arduino motor driver shield, with a servo motor, an ultrasonic sensor, Bluetooth module and a smartphone to give commands to the car. When uploaded this code lets the car run autonomously without any collison. This code can easily be changed according to the purpose and new commands can be added.